Fix pylxd hard dependencies
Since introduction of the LXD/LXC first-party manila driver [1],
generating configuration and running unit tests fail unless pylxd is
installed.  This hard depency is likely unwanted as it is problematic
for distributions that do not include LXD and which will not therefore
distribute a packaged version of pylxd.

This commit converts the pylxd dependency to a soft dependency.  Config
generation succeeds and lxd driver unit tests are skipped if pylxd
cannot be imported.

Additionally, we handle the pylxd exceptions encountered using
pylxd 2.x, as reported here [1].

[1] https://review.openstack.org/#/c/289899/

Change-Id: I8436de2fd994fc73f81f6f3a43c658c874550c10
Closes-bug: #1553280
Closes-bug: #1553208