Squash consequent DB calls in create_share_instance
Currently, when a share instance is created, the 'status'
attribute is updated to 'available'. A check is then
performed to verify if the share instance belongs to
a share that should support replication; if it does,
the 'replica_state' attribute is set to 'active'.

These DB update calls are essentially altering the same
resource. They should be squashed in order to prevent
failures on scripts that probably poll for 'available'
status on the share and immediately look for replicas
of the share.

Change-Id: I36199166fdf3a6a5c115a7140828db896b60562d
Closes-Bug: #1554183