Fix status update for replicas
When an error is raised by a driver during
the update_replica_state loop in the
manager, it will break out of the loop
and not update the other replicas. This change
has the manager set that replica's "replica_state"
and "status" to ERROR instead of raising an exception,
then continue checking the other replicas.

Closes-Bug: 1554270
Change-Id: I6df8d0a4c5f3a7304d51ae97f710f4faf61a5adb