Disallow scheduling multiple replicas on a given pool
Currently, while scheduling a new replica, the ShareReplicationFilter
passes any host that shares the replication_domain of the active
replica's host. On successive runs on the CI, it was observed
that the same host was chosen multiple times. In a real
deployment, there is no practicality in allowing replication
within the same pool.

Pass more information to assist the filter in disqualifying hosts that
already host replicas for the share.

Also treat NoValidHost exception differently from other scheduler
exceptions to be consistent with other scheduling failures.

Change-Id: Iedc7b752c7fc477d6257d100e2277378d59f1fde
Closes-Bug: #1554282