Fix ZFSonLinux driver share replica SSHing
ZFSonLinux share driver uses username and service IP address
for replication actions. And if we change those values in config
and restart manila-share services driver still will use old values
in "promote" and "update/resync" operations.
So, update that info in each "ensure_share" call that is performed
on start of service.

Change-Id: I565d9e41ff788331af5f30d0edcc7f33d05b806b
Closes-Bug: #1552320