update quota of origin user on share extend/shrink
When extending or shrinking a share, the quota of the current context
user gets increased. But when the share gets removed only the quota
of the user that created the share will be decreased.
This patch only updates the quota values for the user, who created the
share, on share extend/shrink actions, to properly decrease quota, once
the share gets deleted.

Change-Id: Iec4b64e96bd0bd3e354532716c9debbce815520d
Closes-Bug: #1555604