Windows driver: fix share access actions
The Windows SMB driver is based on the Generic share driver. In order
to remain consistent and avoid duplicated code, it inherits it,
reusing common code.

The Generic share driver has transitioned to the new 'update_access'
workflow, so the old 'allow_access' and 'deny_access' methods have
been removed, breaking the Windows SMB driver.

This patch fixes this issue by simply readding them in the Windows
driver. Note that this is just a temporary fix, until we move to
the new access rule workflow.

Closes-Bug: #1556261

Change-Id: Iafa0d23e93b1f3d6d2b3c30535c3fc0a1236cdc0