NetApp cDOT driver should honor reserved percentage
There is a common manila.conf parameter 'reserved_share_percentage'
that may be set for each backend, and (if desired) it is the
driver's responsibility to read this value and report it to the
scheduler for each pool. The NetApp cDOT driver has never honored
this value, but it should.  This commit makes this change, and it
removes a couple extraneous capacity fields from the share stats.

Change-Id: Ice52926033e5fd1fd787b71b0e98adccc1a2cbdc
Closes-Bug: #1561704
(cherry picked from commit f56d2fda973cfa3d45e19f18ea2a057b8a9438fd)