Fix Share status when driver migrates
If a driver implements optimized migration, share status is stuck
'migrating' and task state is stuck 'migration_in_progress' if
2-phase migration is not used. This change addresses it by properly
setting the task state and status for this specific scenario.
This change also adds 'host' to model update performed by manager,
to prevent confusion related to drivers having to return such
value in their model update or not.

Change-Id: I4b2c1b02771d8bbb6a8d3ca2719186ad41e5d9af
Closes-bug: #1563823