Developer Reference: Adopt the openstackdocstheme
To allow for our developer reference to be consistent with
the rest of the openstack documentation, we need to import
and render this theme on our developer reference.

This gives users a consistent experience as they may
be acquainted with the rest of the OpenStack documentation
(Admin Guide, User Guide, API Guide etc.)
before perusing the Manila developer reference.

It also provides our devref the "log-a-bug" feature so
readers may bring any mis-information to our attention.

This patch is part of a series of commits to improve
in-tree documentation.

Change-Id: I0e4686a273c5bf45c1d57764cd4c312870b06d6e
Partially-implements: bp improve-manila-developer-docs