Remove retry logic from manage API
Lets say, we try to manage a share and share for some reason
goes in 'manage_error' state. Now, if we try manage command again
with the same parameters as earlier, then the logic in API,
tries to reuse old record in DB.

That is unsettling with the REST way of doing things.

Share Manage API is admin-facing by default, Admin should
delete the share before attempting to manage again.

Hence dropping the retry logic.

APIImpact

Change-Id: I0eee6e108fc263e0d2e092eeaa6d0cba47b53dc4
Closes-Bug: #1566504