HDS_HNAS: Fix improper error message
When attempting to delete a snapshot, if the filesystem is
unmounted, the error message is the output of the delete command.
The delete command should not be invoked in such scenario, and
a proper message should be included in the exception.

(cherry-picked from commit 8b1de88df1d6a84f8f674f3565a1636ffe0a3df3)
Change-Id: If60f8644d4aebf5f364de44c6bef8085d2e539f7
Closes-bug: #1582747