Fix Manila HNAS driver managing a share twice
Manila HNAS driver splits the path to manage a share by ':', so when
passing a path with two colons the last part is ignored, it causes the
possibility to manage a share twice. For example:

manila manage --name test1 --share-type hds ubuntu@hnas#Hitachi nfs
172.24.49.26:/shares/f5b21612-9841-42c8-bec4-ca32ba70a9ae

and

manila manage --name test2 --share-type hds ubuntu@hnas#Hitachi nfs
172.24.49.26:/shares/f5b21612-9841-42c8-bec4-ca32ba70a9ae:1

Changing the split to search for ':/shares/' prevents a share to be
managed twice.

Change-Id: Ie97f749d7093db69a757db7c04041f6f649c7ce9
Closes-Bug: #1613683