Fix race condition updating routes
Generic driver updates routes on host according to the tap
interfaces created for use by share servers. That code
should be within a lock to prevent a thread failing to delete
a route previously listed and already deleted by another thread.

This change moves an existing lock to the function that
performs route changes while also covering the function
that had the lock before.

Closes-bug: #1617461

Change-Id: I798e5d9b0bc9ebefc8661d0ade720c60303cf484