Fix huawei driver username/password encoding bug
While huawei driver reads username/password configuration from
configuration file, if the value is plain text, huawei driver'll
encode this text and write back to configuration file.

huawei driver leverages base64.b64encode to implement the encoding.
The encoded result is a byte string, and then is converted to
unicode string.

In py2, this works fine. However, in py3, the unicode string
converted from byte string is like "b'***'", *** is the orginal
string supposed to be.

So the encoded string is broken, and exception occurs while decoding
from it.

This patch fixes the incorrect unicode string converting and make
the encoded result string right.

Change-Id: I2483ae4999484b3a67c7d443f9b3cc214a3473b2
Closes-Bug: #1613242