Compare the encoded tag more accurately for huawei driver
huawei driver uses a tag '!$$$' to indicate if username/password
is encoded or not in huawei configuration file.

Currently 'find' method is used to check if this tag is included
in the configuration string, but it'll misread if the tag is
included not in front of the string.

So change the comparing statement to a more accurate way, compare
the front slice of the string directly to the tag.

Change-Id: Ic343970578477362460340b01e5766e03c7d63a4
Closes-Bug: 1635073