Fix share manage tempest test cleanup
test_share_manage.ManageNFSShareTest might leave undeleted shares on
backend if some tests are not run. This happens because the shares are
created and unmanaged on resource_setup and are expected to be managed
and deleted on the tests. If the tests don't run, the cleanup will not
delete the shares because they are already unmanaged, thus leaving them
created on the backend.

This patch fixes this issue by moving the share creation and unmanage
operations from resource_setup to the tests themselves.

Closes-bug: #1639848

Change-Id: I2d1790f0a323fdc7b11805842985ca5907dd23c3