[Tempest] Port remote_client into Manila
Manila tempest tests make use of remote_client [1], which won't be making it
to the tempest stable interfaces, as it imports tempest code which would
result in a circular dependency.

This commit ports a reduced version of remote_client into manila code in
order to have manila's tempest plugin to drop the dependency on it.

[1] https://github.com/openstack/tempest/blob/master/tempest/common/utils/linux/remote_client.py

Partially-Implements: bp/tempest-no-deps

Change-Id: I97a8c57adce9cd541766cc1a2f21ca9ceb92efe9