Mark 'v1' API deprecated in the versions response
We announced the deprecation of the v1 API on
the mailing list during the mitaka release. While we
continue to support it in code and run CI tests on it,
we should allow the consumers of the versions API to
acknowledge this change.

This patch does not remove the API. It only marks the
version as "DEPRECATED".

[1] http://lists.openstack.org/pipermail/openstack-dev/2016-June/096517.html

Change-Id: If80bddb68a5b92b604ef582fe07eef72c47d6f2f