Fix setup of DHSS=False mode for generic driver
By moving creation of Nova VMs to later step when Nova configuration
is fully completed. In our case, after merge of [1] we are able
to create Nova VMs only in "test-config" section because latest Nova
step that is required for creation of VMs is performed in the end
of "post-extra" section.

[1] I8742071b55f018f864f5a382de20075a5b444a79

Change-Id: Ib941a5551fe340b39719e3c0e701067ce83bac02
Closes-Bug: #1660304