Fix HNAS driver always handling mountable snapshots
HNAS driver always handles snapshots as if mount_snaphot_support
is always True.
The driver always create exports in snapshot_create and tries to
delete it in snapshot_delete. Also, the driver tries to add rules
for exports in snapshot_update_access without checking if the export
exists.

Fixing it by checking the key 'mount_snapshot_support' in shares and
checking if the snapshot has an export before adding or removing
rules.

Change-Id: If25d68d04400f0e9015f6ad6aff845f9c5d47f94
Closes-bug: #1660420
Closes-bug: #1660421