Update share server provisioning for share groups
The _provide_share_server_for_share_group method derived from the
corresponding _provide_share_server_for_share method, but differs
from it as well.

Update the parameter check because the group provisioning method
always requires that a share_network_id be passed, even when a
snapshot parameter is supplied.  Also update the docstring for
this method to apply to share groups rather than shares and to
be more accurate.

TrivialFix

Change-Id: Ib39c4cb95fcc833905e82213cc3deaef86ab7d07