CI: Update tempest commit
Manila pins the version of tempest it tests on the gate.
This is antithetical to the tempest plugin mechanism. So,
our gate breaks when some plugin relies on code from tempest
that is not part of the commit manila pins in the CI.

While we are working towards a permanent fix to this problem,
let's update our pin as a workaround.

Closes-Bug: #1693832

Change-Id: Ide131229efdea5a48dc389449602f2b0abb699f7