NetApp: Define 'preferred' to False instead of none
Define preferred to false instead of none in the NetApp driver
as setting it to none will trigger a traceback later on because
the database doesn't allow null values to be inserted.
"preferred" export location metadata must be set to true or
false.  warning logging is also added to the SQLAlchemy API
if a none value is used to update the export location metadata.

Change-Id: Ie7420c9286cc42eaa1361eeffe607b9b5b6b3fbd
Closes-Bug: 1703660