doc migration: install guide
There is little material on manila in the centralized Install
Guide to migrate as outlined in the migration spec [1], so copy
from our local install guide.  After we complete this migration,
we can remove the job that builds the local install guide and
remove it from the manila tree.

[1] https://specs.openstack.org/openstack/docs-specs/specs/pike/os-manuals-migration.html

Change-Id: Ibe3588c3f4560c037cf109058fc357234e70a846
Partial-Bug: #1706181
Needed-By: I04237021943bb7501acb9cfb7252be2cbf07ac4b
Depends-On: I7924d94b82e7c8d9716bad7a219fc38c57970773
Depends-On: Ia750cb049c0f53a234ea70ce1f2bbbb7a2aa9454