doc migration: configuration reference
In accord with the OpenStack documentation migration spec [1],
this patch pulls over material from the centralized Configuration
Reference so that it will not be lost.  As indicated in the spec,
dynamic generation of this material will be left for later work.

[1] https://specs.openstack.org/openstack/docs-specs/specs/pike/os-manuals-migration.html

Closes-Bug: #1706181
Depends-On: Ib0fca0c16e14c05307bfc31ae50c65271b7402f1
Depends-On: Ia750cb049c0f53a234ea70ce1f2bbbb7a2aa9454

Change-Id: I7f3b584ac9907da1f07dedd26324cf4703bc7e6c
(cherry picked from commit fae499cb7347e8bca1063f96334356282d1fe703)