[Trivialfix] Remove the useless parameter 'ext_mgr'
The parameter 'ext_mgr' is not used in method _setup_routes() in
subclasses:
manila.api.v2/router.APIRouter
manila.api.v1.router.APIRouter
manila.api.versions.VersionsRouter
So it should be removed to make the code clean.

The superclass of them is manila.api.openstack.__init__.APIRouter,
the 'ext_mgr' is removed from method _setup_routers() in superclass.

In method__init__() in superclass, the _setup_routers() is called,
so 'ext_mgr'is removed in related calling statement.

Change-Id: I25e2d5447b328c88c660ae17025f65719290b21b