Add default configuration files to data_files
In order to make it simpler to use the default configuration files
when deploying services from source, the files are added to pbr's
data_files section so that the files are included in the built wheels
and therefore deployed with the code. Packaging and deployment tools
can then more easily use the default files if they wish to.

This pattern is already established with similar
files for neutron and the glance metadefs as has
been mentioned in the related bug report.

Change-Id: I4c42caf57f3bb27f92e045eb63b7580789402207
Closes-Bug: #1718356