doc: move stuff from contributor to admin
As agreed in weekly manila community meetings and at the
Queens PTG, several documents currently presented to contributors
are really of general interest to cloud administrators and should
be presented accordingly.

Move the docs on manila back ends, the feature support mapping,
capabilities and extra specs, and export locations and metatdata
to the admin section.

Change-Id: I6d9dec56890fb3a9bc0539c45c43233517cea048