Fix drivers_private_data update on deleted entries
The drivers_private_data is expected to be able to
save new entries using the same keys as previously deleted
entries.

Before performing an update, it retrieves all entries
related to the same entity to check if it needs to
perform an update (including on deleted entries) or a
new insertion.

Unfortunately, the misinterpretation of a parameter in the
manila DB API led to a bug that prevented deleted entries
to be retrieved.

This bugfix replaces the query ensuring that all entries
(active and deleted) are retrieved so the update validation
works.

Closes-bug: #1721123

Change-Id: Id94544c9b8a534d7f0fe4e16bc57546002e33373
(cherry picked from commit ebac41d6314adb4046c097fcb1046308f3776d31)