Add functional jobs to check and gate pipelines
This commit brings functional test jobs to stable/ocata.
It began as a backport of 6fec3f23566a6952215ef1e7250a9ea9ecd57f93
however, we'll run a minimum set of jobs
on this branch because it is transitioning to
extended maintenance.

This commit also reverts the old LVM centos-7 job to
running on Ubuntu Xenial Xerus since there are
outstanding package issues with centos that
are out of the scope of the manila team. See [1]

Check Queue:
  - Voting:
    - manila-tox-genconfig
    - manila-tempest-minimal-dsvm-dummy
    - manila-tempest-minimal-dsvm-lvm
  - Non-voting:
    - openstack-tox-cover
    - openstack-tox-pylint
    - manila-tempest-dsvm-mysql-generic
    - manila-tempest-dsvm-postgres-container
    - manila-tempest-dsvm-postgres-zfsonlinux
    - manila-tempest-dsvm-postgres-generic-singlebackend
    - manila-tempest-dsvm-generic-no-share-servers
    - manila-tempest-dsvm-scenario

Gate Queue:
  - Voting:
    - manila-tox-genconfig
    - manila-tempest-minimal-dsvm-dummy
    - manila-tempest-minimal-dsvm-lvm

[1] https://bugs.launchpad.net/openstack-requirements/+bug/1778971

Depends-On: https://review.openstack.org/#/c/604657/
Change-Id: I7c9fbbb22860a2d246a9e6cb834f7c87875344f0