Migrating legacy jobs
Moving legacy jobs from openstack-zuul-jobs to manila
repository. In order for us to start reworking the legacy
jobs to native v3 jobs, we need to do so.
This is the first step as described in the manual
and will also make it easier to remove the in-tree
tempest plugin sooner.

Change-Id: I7c9fbbb22860a2d246a9e6cb834f7c87875344f0