Migrating legacy jobs
Moving legacy jobs from openstack-zuul-jobs to manila
repository. In order for us to start reworking the legacy
jobs to native v3 jobs, we need to do so.
This is the first step as described in the manual
and will also make it easier to remove the in-tree
tempest plugin sooner.

Depends-On: Ib83267ef75f91fb8f95af312d644ebe19c6307c9
Depends-On: Ia9226329035508c37ca499e6efbd2382502eb141
Change-Id: I7c9fbbb22860a2d246a9e6cb834f7c87875344f0
(cherry picked from commit 6fec3f23566a6952215ef1e7250a9ea9ecd57f93)