Impove coverage job accuracy
by informing it that we use eventlet for concurrency.

We have many false negative coverage runs, even for changes that
do not touch the code under test, for example in doc only change [1].
According to the coverage documentation, unless we indicate that
we are using an alternate to the thread library for concurrency
we will get "very wrong results." [2]

Set this option correctly.  Also avoid false precision in the
coverage report.

[1] http://logs.openstack.org/01/514901/2/check/openstack-tox-cover/6df6369/

[2] https://coverage.readthedocs.io/en/coverage-4.4.1/config.html

TrivialFix

Change-Id: I7398a45cdab8aa4191a55cbb86b9e6499b09f021
(cherry picked from commit c22c073734d26398a3608e23d1ddffd015805342)