Fix issue with different decimal separators
An issue was discovered in the ZFSonLinux driver wherein on a system
set to have a locale where the decimal separator is not a period but
rather a comma the method translate_string_size_to_float() would fail
because it's regex to find sizes in strings did not account for any
other decimal separators other than periods.

This fix updates the regular expression to accept either a period or
a comma as a decimal separator in a size string.

Many thanks to Dr. Clemens Hardewig for reporting the original bug
and providing an updated regular expression.

Change-Id: I15da4aaff90814eed03816db013d6acc89ba2ee8
Closes-Bug: #1714691
(cherry picked from commit 8800e54f770c4c4a6f813a5ea72bb12316482d87)