Remove hdfs job from check queue
In the manila weekly community meeting it was resolved [1]
to remove unmaintained drivers from the check queue to avoid
waste of infrastructure resources and to remove "red noise"
from the gerrit review pages.

Do so for the hdfs driver.  Leave the job defined as it is run
in devstack-hdfs-plugin.

[1] http://eavesdrop.openstack.org/irclogs/%23openstack-meeting-alt/%23openstack-meeting-alt.2017-11-02.log.html#t2017-11-02T15:26:17

Change-Id: Ifdd55572e9a2ba243fe074476d9459acaf474f54
(cherry picked from commit a30085cca30e35febfed6efb22f5066502110577)