Remove glusterfs-native job from check queue
In the manila community meeting it was resolved [1]
to remove unmaintained drivers from the check queue
to avoid waste of infrastructure resources and to
remove "red noise" from the gerrit review pages.

Do so for the glusterfs-native driver.  Leave the
job defined as it is run in the experimental queue
and in devstack-glusterfs-plugin.

[1[ http://eavesdrop.openstack.org/irclogs/%23openstack-meeting-alt/%23openstack-meeting-alt.2017-11-02.log.html#t2017-11-02T15:26:17

Depends-On: I33fa0026670724cb1d7ca047b4948e4110078342
Change-Id: I87f15ee7f73a8b4a16ae0549d0f2fae4b4c23ec2
(cherry picked from commit 4b3172be57e39d7ca3a4be19ccca582ab8a583c6)