Fix boolean types in db migration tests
In gate py27 and py35 unit tests are failing on
alembic migration tests with exceptions like this [1].
Apparently sqlAlchemy 1.2.1 enforces type checks for
database fields that 1.1.15 just coerced.

Change strings to booleans and ints in the test
code so that fields in the "fakes" match the
fields in the sqlalchemy models.

Co-Authored-By: Tom Barron <tpb@dyncloud.net>

[1] http://logs.openstack.org/68/539468/1/check/openstack-tox-py27/db4a572/job-output.txt.gz#_2018-01-31_09_26_49_316004

Change-Id: Iadffb8a06b4d948d77fc00b85118e538b0a19b24