Create a generic share type (api tests)
Tempest assumes a default share type already created
to work. This means, if a default share type is not
created and not specified in the conf file, tempest
tests fail. A workaround is to create a share type
as part of the environment setup for all the tests
that need it. This patch set does that.

(cherry picked from commit I15880e400df30918762ebd7304244b4a27200168)
Closes-Bug: #1743472

Change-Id: Icaa02617a4a3dbd14fdbfa2ecf5d44aac497f60e