Fix LVM driver not handling IPv6 in recovery mode
Recovery mode is currently used by host-assisted share
migration to bulk remove and add back rules while casting
them to read only.

Unfortunately, it is not handling IPv6 addresses when removing
the rules in this code path. This patch adds a call to a function
that handles the IPv6 addresses, adding brackets around them.

Change-Id: Icba4b22ad87ccfabcc02078648b1abc6410e5353
Closes-bug: #1746723