Fix manila logging rabbitmq password in debug mode
Manila will display the rabbitmq password if debugging is enabled.
This patch will ensure that the rabbitmq password is no longer
displayed in the connection log for Manila when debugging is
enabled by looking for the rabbitmq key and not printing it.

There should likely be an effort to utilize Oslo's secret flag for
options to truly fix this issue for this and other sensitive options.

Change-Id: I97cc88354d9b54057350c70c4742055197540d1a
Closes-Bug: 1750074
(cherry picked from commit 05e4f14ea1fbc7ce61c8f57a082080d09c63e357)