Log config options with oslo.config
This removes some custom code inherited from Cinder which was handling
the output of secret options in a bad way. This patch utilizes Oslo's
existing utilities to output the Manila configuration options securely.

Filtering will be done with the "secret=True" option flag.

Major thanks to Eric Harney for introducing this fix to Cinder.

Change-Id: I894e011680661c0b73b9592f70a6457e403f18c6
Related-Bug: #1750074