Use 'Default' as the value of domain name in install guide
The Manila install guide use default in some place as the value of
user_domain_name and project_domain_name, whereas the Keystone
install guide, which people would be consulting earlier, instructs
the installer to use 'Default'.  Since the name is case-sensitive,
this is causing problems for first-time installers of OpenStack.

Ref: https://bugs.launchpad.net/keystone/+bug/1229093

Change-Id: I1cf4345abd3ae8923f8d1fbd0706737494745d84
Closes-bug: #1766475