Limit formatting routes when adding resources
By default, routes.mapper.Mapper.resource adds a bunch of formatted
routes that accept anything after a '.'. Our spec says only .xml
and .json are accepted so limit the formatting to those valuse.

This allows identifiers with a '.' in urls. A few tests were added
to the extensions test to prove that .xml and .json are stripped
but other values are not.

It is already modified in cinder [1].

[1] https://bugs.launchpad.net/cinder/+bug/1307956

Change-Id: I91d8aa967446b5d7ce52f1b245149304494777f3
Closes-bug: #1774353
(cherry picked from commit 3c37b942e73937c854b6ffa0e1e15164673ae1f1)