Adds export path option to Quobyte driver
A new option for configuring the export path of
shares provided by the Quobyte driver is added.
This allows adapting the export path to the settings
of the Quobyte NFS servers 'Pseudo' option.

Besides the new option this also fixes a minor
coding style issue regarding a mutable default
argument in jsonrpc.py.

Implements: blueprint qb-export-path-option
Closes-Bug: #1773929

Change-Id: Ibd486c8418469045a0988ef66a1c5cef810d3eae