Remove confusing DB deprecation messages
When share instances were introduced a number of fields
were moved in the database model to the underlying share instance
and proxied back to the share in order to maintain compatability
with APIs that assume these to belong to the share.  A deprecation
warning was added for these.  The warning spams the API log and
confuses end users because it sounds like they or their client
should make a change to get rid of them and that cannot actually
be done.  The API would have to change to expose share instances
to end users rather than just administrators and that is not
something that we plan to do.

Remove these confusing warnings.

Closes-Bug: #1776947

Depends-On: https://review.openstack.org/578558
Change-Id: I95e99b26f416bbf25cfd1c8f7295f3ba6220efb3
(cherry picked from commit 8bdbbd86ae8efe05950ad0651c40aa17c8ec71c7)
(cherry picked from commit 0c182bca8671dff0296899eef05ac0c473b1a115)