NetApp ONTAP: Fix use of multiple subnets with DHSS=True
NetApp ONTAP Multi-SVM driver was raising an error while trying to
create shares on multiple subnets that belong to the same neutron
network, as it was trying to map multiple ipspaces to the same VLAN
port.

This fix allows the driver to use the same ipspace and VLAN port across
all subnets belonging to the same neutron network.

Change-Id: If9cbb34a890ee44806c404085e40cc924a1296a7
Closes-Bug: #1774159