INFINIDAT: unit tests - remove fake exception body
The class FakeInfinisdkException is used in the tests code to mock the
exceptions raised by infinisdk (our internal library for connecting to
the InfiniBox REST API). In earlier versions, this exception was
implemented by inhereting from the base Exception class and
implementing its constructor, however, this code is redundant,
and basic inherithance should do.

Change-Id: I44059d41143eaee6d7dcb073e21c9c840dacae3e